+++ app/Http/Controllers/PeriodController.php
@@ -10,7 +10,8 @@ class PeriodController extends Controller
 {
   public function index()
   {
-    $periods = Period::get();
+    $login_id = \Auth::id();
+    $periods = Period::where('user_id', $login_id)->get();
 
     return compact('periods');
   }
@@ -36,19 +37,21 @@ class PeriodController extends Controller
 
   public function update(Request $request)
   {
+    $login_id = \Auth::id();
     $periods = $request->input('periods');
     $delete_periods = $request->input('delete_periods');
 
     foreach($delete_periods as $delete_period){
-      Period::where('period_id', $delete_period['period_id'])->delete();
+      Period::where('user_id', $login_id)->where('period_id', $delete_period['period_id'])->delete();
     }
 
     foreach($periods as $period){
       if(array_key_exists('is_new', $period)){
         unset($period['is_new']);
+        $period['user_id'] = $login_id;
         Period::create($period);
       }else{
-        Period::where('period_id', $period['period_id'])->update($period);
+        Period::where('user_id', $login_id)->where('period_id', $period['period_id'])->update($period);
       }
     }
